@@ -3,18 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@types/jest": "29.5.12",
-    "@types/node": "20.11.17",
-    "@types/ramda": "^0.29.10",
-    "@types/react": "^18.2.55",
-    "@types/react-dom": "18.2.19",
-    "@types/react-lazyload": "^3.2.3",
-    "@types/react-loadable": "^5.5.0",
-    "@types/react-router-dom": "^5.3.3",
     "hyphenation.fi": "^0.2.1",
     "hypher": "^0.2.5",
     "moment": "^2.24.0",
-    "node-sass": "^9.0.0",
     "ramda": "^0.29.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -25,17 +16,13 @@
     "react-markdown": "^9.0.1",
     "react-modal": "^3.13.1",
     "react-router-dom": "^5.1.0",
-    "react-scripts": "^5.0.1",
     "react-youtube": "^10.1.0",
-    "socket.io-client": "^2.5.0",
-    "typescript": "5.1.6",
-    "yarn": "^1.22.4"
+    "socket.io-client": "^2.5.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "webpack serve --mode development --config ./config/webpack.dev.js",
+    "build": "webpack build",
+    "test": "jest test"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -48,10 +35,34 @@
   ],
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@types/react-helmet": "^6.1.0"
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.16",
+    "@swc/core": "^1.11.18",
+    "@types/jest": "29.5.12",
+    "@types/node": "20.11.17",
+    "@types/ramda": "^0.29.10",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "18.2.19",
+    "@types/react-helmet": "^6.1.0",
+    "@types/react-lazyload": "^3.2.3",
+    "@types/react-loadable": "^5.5.0",
+    "@types/react-router-dom": "^5.3.3",
+    "css-loader": "^7.1.2",
+    "html-webpack-plugin": "^5.6.3",
+    "react-refresh": "^0.17.0",
+    "resolve-url-loader": "^5.0.0",
+    "sass": "^1.86.3",
+    "sass-loader": "^16.0.5",
+    "style-loader": "^4.0.0",
+    "swc-loader": "^0.2.6",
+    "typescript": "5.1.6",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.99.5",
+    "webpack-cli": "^6.0.1",
+    "webpack-dev-server": "^5.2.1",
+    "yarn": "^1.22.4"
   },
   "volta": {
-    "node": "20.18.3",
+    "node": "22.14.0",
     "yarn": "1.22.22"
   }
 }
