@@ -26,7 +26,7 @@ const friendsInfo = [
     title: 'Tekniikan Akateemiset',
     url: 'https://www.tek.fi/fi',
     logo: (
-      <g className="f1" stroke-width=".20795">
+      <g className="f1" strokeWidth=".20795">
       <path d="m368.91 89.369a6.8063 6.8063 0 0 1-4.0177-1.3101l-49.553-36.088a6.8334 6.8334 0 0 1-0.24954-10.864l49.56-39.611a6.8355 6.8355 0 0 1 8.5261 10.678l-42.554 34.019 42.321 30.815a6.8355 6.8355 0 0 1-4.0302 12.361"/>
       <path d="m261.62 13.67h-150.42a6.8355 6.8355 0 1 1 0-13.669h150.42a6.8355 6.8355 0 1 1 0 13.669"/>
       <path d="m165.82 89.619c-3.7764 0-6.8355-2.9322-6.8355-6.5464v-38.942c0-3.6142 3.059-6.5443 6.8355-6.5443s6.8355 2.9301 6.8355 6.5443v38.942c0 3.6142-3.0611 6.5464-6.8355 6.5464"/>
@@ -125,7 +125,7 @@ const friendsInfo = [
     title: 'Fonum',
     url: 'https://www.fonum.fi',
     logo: (
-      <path className='f1' d="m100.01 149.83c-7.9511-1.2397-14.684-6.1199-18.263-13.237-0.90876-1.8074-1.8992-4.8619-2.2705-7.0022-0.29174-1.6817-0.30613-4.245-0.30613-54.535 0-50.29 0.01439-52.853 0.30613-54.535 1.8544-10.689 10.049-18.69 20.741-20.251 1.596-0.233 4.8275-0.25772 27.438-0.20988l25.629 0.054221 1.63 0.37347c2.286 0.52377 4.3024 1.26 6.3195 2.3073 6.5242 3.3875 11.037 9.3562 12.679 16.77l0.3792 1.7122v107.56l-0.39059 1.7445c-0.53065 2.37-1.2315 4.3095-2.3029 6.373-3.3725 6.4953-9.2665 10.972-16.684 12.672l-1.63 0.37354-25.881 0.0344c-21.509 0.0285-26.136-5e-3 -27.393-0.20148zm27.735-4.4572c2.496-0.46142 4.2592-2.0948 4.8809-4.5218 0.69365-2.7076-0.71978-5.6608-3.2902-6.8746-0.80305-0.37921-1.1448-0.43647-2.6049-0.43647-1.5925 0-1.7358 0.03-2.734 0.57124-1.2604 0.6835-2.1057 1.5752-2.705 2.8538-0.37549 0.80101-0.44277 1.1742-0.44286 2.457-9e-5 1.2868 0.0666 1.6549 0.44699 2.4663 0.87077 1.8575 2.5466 3.1719 4.4519 3.4916 0.44826 0.0752 0.85784 0.14565 0.91019 0.15652 0.0523 0.0109 0.54151-0.0627 1.087-0.16359zm34.026-74.015v-55.544l-69.996 0.08482-0.04227 55.501-0.04228 55.501h70.081zm-158.31-1.0084v-57.56l65.459 0.08488 0.08854 11.175-53.867 0.08505v38.653l48.489 0.08519-0.04432 5.5033-0.04432 5.5033-48.485 0.08519v53.945h-11.596zm317.67 57.391c-4.5514-0.26974-9.1826-0.95027-12.57-1.8471-5.922-1.5678-11.439-4.9235-14.998-9.1226-4.7163-5.5641-7.2437-12.341-8.0993-21.717-0.10727-1.1755-0.16875-16.49-0.16884-42.059l-1.5e-4 -40.21 11.344 0.0885 0.0859 40.67c0.0958 45.354 0.0105 42.17 1.2597 47.064 1.6018 6.2748 4.8366 10.611 9.7462 13.063 2.3119 1.155 5.8644 2.0578 10.168 2.5841 3.4564 0.42272 11.707 0.4241 15.125 3e-3 7.3271-0.90372 11.586-2.6212 14.951-6.029 2.9619-2.9998 4.7489-7.1596 5.8327-13.578 0.27786-1.6455 0.30398-4.5334 0.38675-42.771l0.0888-41.006h11.428v40.418c0 37.703-0.0201 40.593-0.29985 43.023-0.99547 8.6478-3.4062 14.896-7.7873 20.183-3.6905 4.4538-9.1272 7.8014-15.287 9.4127-5.6017 1.4654-14.46 2.2299-21.205 1.8302zm-131.59-57.769 0.0423-57.098h8.2348l29.578 44.405c16.268 24.423 29.637 44.407 29.709 44.409 0.072 2e-3 0.0913-2.1323 0.043-4.7435-0.0483-2.6112-0.12856-22.614-0.17838-44.451l-0.0905-39.704h11.775v114.28l-8.3189-0.01-29.564-44.362c-16.26-24.399-29.592-44.362-29.625-44.362-0.0916 0 0.11137 86.281 0.20598 87.6l0.0814 1.1344h-11.935zm192.76 0 0.0422-57.098h8.7184l19.158 41.213c10.537 22.667 19.188 41.183 19.225 41.146 0.0367-0.03671 8.6738-18.582 19.194-41.213l19.127-41.146 8.7742-0.08962v114.28h-11.815l0.0904-40.18c0.0497-22.099 0.0314-40.117-0.0409-40.04-0.0722 0.07733-7.0755 14.869-15.563 32.87l-15.431 32.729-8.75-8e-3 -30.923-65.598-0.0846 80.226h-11.764z" stroke-width=".52518"/>
+      <path className='f1' d="m100.01 149.83c-7.9511-1.2397-14.684-6.1199-18.263-13.237-0.90876-1.8074-1.8992-4.8619-2.2705-7.0022-0.29174-1.6817-0.30613-4.245-0.30613-54.535 0-50.29 0.01439-52.853 0.30613-54.535 1.8544-10.689 10.049-18.69 20.741-20.251 1.596-0.233 4.8275-0.25772 27.438-0.20988l25.629 0.054221 1.63 0.37347c2.286 0.52377 4.3024 1.26 6.3195 2.3073 6.5242 3.3875 11.037 9.3562 12.679 16.77l0.3792 1.7122v107.56l-0.39059 1.7445c-0.53065 2.37-1.2315 4.3095-2.3029 6.373-3.3725 6.4953-9.2665 10.972-16.684 12.672l-1.63 0.37354-25.881 0.0344c-21.509 0.0285-26.136-5e-3 -27.393-0.20148zm27.735-4.4572c2.496-0.46142 4.2592-2.0948 4.8809-4.5218 0.69365-2.7076-0.71978-5.6608-3.2902-6.8746-0.80305-0.37921-1.1448-0.43647-2.6049-0.43647-1.5925 0-1.7358 0.03-2.734 0.57124-1.2604 0.6835-2.1057 1.5752-2.705 2.8538-0.37549 0.80101-0.44277 1.1742-0.44286 2.457-9e-5 1.2868 0.0666 1.6549 0.44699 2.4663 0.87077 1.8575 2.5466 3.1719 4.4519 3.4916 0.44826 0.0752 0.85784 0.14565 0.91019 0.15652 0.0523 0.0109 0.54151-0.0627 1.087-0.16359zm34.026-74.015v-55.544l-69.996 0.08482-0.04227 55.501-0.04228 55.501h70.081zm-158.31-1.0084v-57.56l65.459 0.08488 0.08854 11.175-53.867 0.08505v38.653l48.489 0.08519-0.04432 5.5033-0.04432 5.5033-48.485 0.08519v53.945h-11.596zm317.67 57.391c-4.5514-0.26974-9.1826-0.95027-12.57-1.8471-5.922-1.5678-11.439-4.9235-14.998-9.1226-4.7163-5.5641-7.2437-12.341-8.0993-21.717-0.10727-1.1755-0.16875-16.49-0.16884-42.059l-1.5e-4 -40.21 11.344 0.0885 0.0859 40.67c0.0958 45.354 0.0105 42.17 1.2597 47.064 1.6018 6.2748 4.8366 10.611 9.7462 13.063 2.3119 1.155 5.8644 2.0578 10.168 2.5841 3.4564 0.42272 11.707 0.4241 15.125 3e-3 7.3271-0.90372 11.586-2.6212 14.951-6.029 2.9619-2.9998 4.7489-7.1596 5.8327-13.578 0.27786-1.6455 0.30398-4.5334 0.38675-42.771l0.0888-41.006h11.428v40.418c0 37.703-0.0201 40.593-0.29985 43.023-0.99547 8.6478-3.4062 14.896-7.7873 20.183-3.6905 4.4538-9.1272 7.8014-15.287 9.4127-5.6017 1.4654-14.46 2.2299-21.205 1.8302zm-131.59-57.769 0.0423-57.098h8.2348l29.578 44.405c16.268 24.423 29.637 44.407 29.709 44.409 0.072 2e-3 0.0913-2.1323 0.043-4.7435-0.0483-2.6112-0.12856-22.614-0.17838-44.451l-0.0905-39.704h11.775v114.28l-8.3189-0.01-29.564-44.362c-16.26-24.399-29.592-44.362-29.625-44.362-0.0916 0 0.11137 86.281 0.20598 87.6l0.0814 1.1344h-11.935zm192.76 0 0.0422-57.098h8.7184l19.158 41.213c10.537 22.667 19.188 41.183 19.225 41.146 0.0367-0.03671 8.6738-18.582 19.194-41.213l19.127-41.146 8.7742-0.08962v114.28h-11.815l0.0904-40.18c0.0497-22.099 0.0314-40.117-0.0409-40.04-0.0722 0.07733-7.0755 14.869-15.563 32.87l-15.431 32.729-8.75-8e-3 -30.923-65.598-0.0846 80.226h-11.764z" strokeWidth=".52518"/>
     )
   },
   {
@@ -249,7 +249,7 @@ const friendsInfo = [
     url: 'https://tamppi.fi',
     logo: (
       <g className="f1" transform="matrix(.59 0 0 .59 153 -145)">
-        <path d="m19 484c-.059-.095.17-1 .51-2.1.34-1.1 1-3.4 1.5-5.1.51-1.8 1.1-3.4 1.3-3.6.33-.37.66-.38 7.3-.28 6.4.1 10 .35 17 .98 1.2.12 2.3.27 2.5.34.17.062-.76.33-2.1.61-1.3.27-3.2.68-4.3.9-7 1.4-17 5-22 7.9-1 .59-1.5.71-1.7.41zm384-1.1c-.51-.093-2-.31-3.4-.49-1.4-.18-3.1-.41-3.8-.53-.74-.11-2.7-.35-4.3-.53-1.6-.18-3.6-.4-4.4-.5-.79-.096-2.5-.27-3.8-.39-1.3-.12-3.3-.31-4.5-.43-5.4-.52-11-.7-20-.71-11-.012-12 .072-19 .92-1.9.25-5.3.91-6.9 1.4-.61.17-1.2.23-1.3.14-.095-.095-.096-.63-.00044-1.2.16-.97 1.2-6.4 2.8-15 .15-.79.42-2.5.61-3.7.19-1.2.47-3 .64-3.8.53-2.8 1.5-8.8 1.7-10 .19-1.5.4-2.8 1.2-7.6.22-1.3.55-3.5.74-4.9.18-1.4.57-4.1.85-6 .28-1.9.61-4 .72-4.7.24-1.6 1-7.5 1.2-8.8.054-.51.17-1 .25-1.1.084-.09 1.1-.38 2.2-.65 2.2-.53 5.5-1.7 10-3.5 9.6-3.8 11-4.8 18-13 3.6-4.4 6.3-8.3 7.1-10 .28-.68.72-1.7.97-2.2.99-2 2.1-5.9 3.7-13 .31-1.4.8-3.3 1.1-4.3.94-3.4 1.3-6.6 1.3-13-.005-4.5-.1-6-.6-9.6-.31-2.2-1.5-7.4-2.1-9.5-.29-.91-.53-1.8-.54-2-.021-.78-1.8-5.4-2.6-6.5-1.1-1.7-8.6-9-11-11-1.3-.99-3.2-2.3-4.3-3-5.9-3.6-7-4.3-8.8-5-1.1-.42-2.3-.95-2.8-1.2-.45-.23-2.3-1-4-1.7-1.8-.71-3.3-1.4-3.4-1.6-.13-.17-.29-.74-.36-1.3-.071-.52-.55-2.6-1.1-4.7-.51-2-.93-3.9-.93-4.1 0-.22.21-.48.46-.59.83-.35 11-2.4 15-3 1.2-.18 2.5-.37 2.7-.43.52-.11 7.4-1.1 8.5-1.2.4-.046 2.1-.22 3.8-.4 1.7-.17 4-.41 5-.54 6.2-.71 20-1.1 27-.81 6.9.26 11 .52 13 .83.74.12 2.6.4 4.1.62 4.1.59 4.8.72 8.6 1.6 8 2 15 4.7 20 8.1 5.4 3.5 10 8.7 13 14 .49.91 1.3 2.1 1.9 2.8 2.4 2.6 5.6 7.2 8.8 13 3.5 5.9 6.9 15 8.1 22 .18.96.43 2.3.57 3 .36 1.8.34 12-.026 14-.91 5-1.2 6.1-2.3 9.6-3.2 9.7-11 21-18 25-7.7 5-16 8.1-26 9.1-3 .33-12 .32-15-.021-6.2-.7-13-2.2-17-4-.93-.36-1.7-.54-1.8-.44-.28.29-.49 3.7-1 17-.27 6.7-.27 18-.00086 24 .58 13 .88 17 1.2 19 .11.69.2 1.8.2 2.4 0 .66.091 1.8.2 2.5.11.69.31 2.1.43 3.1.13 1 .31 2.4.41 3 .6 3.9.78 5.2.68 5.4-.13.2-1.3.18-2.5-.05zm7.7-125c4.7-.56 8.5-1.7 11-3.2 1.3-.86 2.8-2.5 3.4-3.7.39-.85.45-1.3.43-3.6-.015-2.2-.098-2.9-.46-3.9-1.4-3.9-4.5-6.1-10-7.4-.85-.19-2.9-.27-6.5-.26-4.4.009-5.7.085-8 .46-3.4.56-9.3 1.9-9.6 2.1-.18.18-.017 2.9.81 13 .16 2 .3 4 .3 4.4.006.4.087.84.18.99.48.76 14 1.1 19 .54zm-332 119c-.98-.36-3.3-1-7.8-2.1-1.8-.46-3.8-.96-4.4-1.1-2.3-.6-8.8-1.9-11-2.3-1.2-.18-2.7-.41-3.3-.51-3.4-.62-4.2-.74-10-1.3-4.1-.44-15-.62-19-.32-6.1.44-13 1.8-17 3.6-1.1.42-1.4.48-1.6.27-.14-.14-.67-1.2-1.2-2.4-.5-1.2-1.4-3.2-2-4.5-.59-1.3-1.5-3.4-2.1-4.6-.54-1.2-1.5-3.4-2.2-4.8-.64-1.4-1.6-3.5-2.1-4.6-.49-1.1-1.7-3.9-2.8-6.2-1-2.3-1.9-4.2-1.9-4.3 0-.044-.32-.76-.7-1.6-.39-.83-1.5-3.4-2.5-5.6-1-2.3-2-4.3-2.1-4.4-.31-.31-.5-.3-2.6.12-2.2.45-3.1.59-7.7 1.2-1.3.18-3 .42-3.7.52-4.7.67-13 1-19 .88-3.9-.092-9.4-.54-11-.89-.51-.11-2.2-.48-3.8-.81-1.6-.33-3.7-.89-4.7-1.2-2.5-.84-2.7-.82-3.2.25-.99 2.1-3.3 9.1-4.1 13-1.6 6.5-2.1 8.7-2.8 14-.18 1.2-.42 2.7-.54 3.4-.28 1.7-.58 4.4-.8 7.4-.18 2.4-.33 3.2-.6 3.2-.073.00078-.97-.18-2-.41-2.9-.64-7.3-1.5-8.7-1.7-3.2-.43-7.1-.84-10-1.1-4.1-.31-21-.3-26 .005-8.3.49-11 .68-11 .78-.23.06-1.2.19-2.3.29-2.2.22-4.9.53-7.2.83-4.1.54-7.7 1-8.1 1.1-.24.067-.49.032-.56-.076-.11-.18.45-1.1 4.6-7.6.87-1.4 2.3-3.6 3.2-5 2.2-3.6 2-3.4 3.7-3.2 3.9.34 10 1.7 15 3.1.87.27 3.6 1.1 6.2 1.7 2.5.68 5.6 1.6 6.9 2 2.9.94 3.6 1.1 3.9.98.48-.19.32-1.2-.43-2.7-1.7-3.3-4.5-11-5.2-13-.18-.76-.55-2.1-.82-3-.27-.87-.88-3.1-1.3-5-.46-1.9-1-3.9-1.3-4.5-.35-.9-1.3-4.5-2.3-9.1-.16-.68-.5-2.2-.75-3.3l-.47-2 .41-.74c.23-.41 1.1-1.9 1.9-3.4 3.9-6.9 5.5-9.8 7.9-14 .7-1.3 1.5-2.8 1.8-3.3.28-.51 1.1-2 1.8-3.3.68-1.3 1.7-3.2 2.2-4.2 3.4-6.4 16-32 16-32 0-.053 1.1-2.3 2.3-5 1.3-2.7 2.8-6 3.4-7.3.59-1.3 1.2-2.4 1.3-2.5.54-.21 7.8.94 12 1.8 3.3.79 8.1 2.1 8.8 2.5l.67.31 1.3-1.2c.73-.65 1.3-1.2 1.3-1.3-.0055-.065-.38-.49-.83-.94-1.4-1.4-3.8-5.5-5.1-8.8-2-5-3.2-8.4-3.8-10-.81-3.1-2-9-2-9.8-.00036-.52 2.6-7.5 3.4-9 .14-.28.46-.54.71-.58.25-.036 1.9.32 3.7.79 13 3.3 22 5.3 31 6.5 1.4.18 3 .41 3.7.53 3.4.57 17 .56 19-.016.4-.12 1-.21 1.3-.21.68.003 2.6-.38 5.9-1.2 2.7-.67 6.8-2.2 9.9-3.6 1.3-.63 2.5-1.1 2.7-1.1.21 0 .43.39.64 1.2.18.65.68 2.4 1.1 3.9.44 1.5 1.4 4.8 2.2 7.3 1.5 5.1 2.5 8.3 5.3 17 1.1 3.3 2.2 6.8 2.5 7.9 2.3 7.3 9.3 27 13 35 .7 1.8 1.7 4.4 2.3 5.9 1.2 3.2 5 13 6.3 16 .53 1.2 1.4 3.3 2 4.6 1.1 2.5 5.8 13 8.3 18 4.7 10 13 25 17 33 4.8 8.1 6.7 11 12 18 .61.89 1.1 1.7 1.1 1.8 0 .27-1 .18-2-.16zm-114-81c2.6-.21 5.1-.62 5.1-.82 0-.07-.42-1.1-.93-2.3-2.5-5.7-4.6-14-6.2-26-.073-.54-.21-.98-.29-.98-.089 0-.42.63-.74 1.4-.88 2.1-3.5 7.3-4.8 9.5-2.6 4.5-6.2 9.4-9.5 13-1.1 1.3-2 2.4-2 2.5-.005.22.26.34 3 1.4 5.1 1.9 11 2.7 16 2.2zm499 81c-.13-.16.05-1.1.53-2.7 1.5-5 3.1-13 3.7-18 .19-1.6.43-3.4.53-4 .1-.57.24-1.9.3-2.9.062-1 .21-2.7.32-3.7.57-5 .57-21-.005-27-.12-1.2-.31-3.5-.43-4.9-.24-3-.46-5-.86-8.5-.3-2.5-.23-2.9.69-3.5 1.3-.84 8.2-8.7 10-11 1.8-2.6 4.6-8.6 5.8-13 .58-1.9 1.3-4.3 1.7-5.3.6-1.7.71-2.4 1.2-7.1.11-1.1.24-2 .31-2 .065-.065 1-.003 2.1.14 1.1.14 3 .38 4.3.54 8.8 1.1 19 1.1 29-.1 6.1-.68 14-2.4 19-4.3 1.7-.57 3.2-1 3.4-1 .48 0 .47.21-.14 2.5-.28 1.1-.88 3.8-1.3 6-1 5.1-1.4 7.6-1.9 13-.12 1.5-.31 3.6-.43 4.8-.26 2.8-.27 18-.005 20 .11.91.3 2.9.42 4.4.12 1.5.3 3.5.4 4.3.095.85.28 2.6.4 3.8.13 1.2.28 2.4.33 2.6.054.17.32 1.7.6 3.3.86 5.1 1.5 8.5 1.9 10 3.6 15 6.3 23 12 37 .46 1.1.84 2.1.84 2.2 0 .23-2.3.029-4.6-.4-.91-.17-2.7-.43-4-.59-1.3-.16-3.4-.44-4.7-.61-8.9-1.2-23-2.1-34-2.1-10-.004-23 .9-29 2.1-.96.18-2.4.46-3.3.61-.85.16-2.1.42-2.8.58-.68.16-1.9.45-2.7.63-.8.18-3.3.92-5.6 1.6-3.1.96-4.2 1.2-4.4 1.1zm-361-1.3c-.99-.18-5.9-.97-7.7-1.2-1.4-.21-4.3-.49-8.4-.83-1.3-.11-2.9-.26-3.5-.34l-1.1-.15-2.4-3.6c-1.3-2-2.6-4-2.8-4.5-.2-.51-.9-1.9-1.6-3.1-.66-1.2-1.7-3.2-2.3-4.4-1.1-2.3-9.4-19-12-25-.93-1.8-3.2-6.3-5.1-10-7-14-10-20-11-20-.37-.43-.85-1.5-3.9-9.2-1.2-2.9-1.9-4.9-1.9-5.4.0051-.4.17-2 .36-3.6.19-1.6.43-3.9.53-5.2.099-1.3.33-4 .52-6 .38-4 .55-6.2.89-11 .12-1.9.32-4.6.43-6 .45-5.3.67-19 .54-33-.072-7.6-.19-14-.26-15l-.13-1.2.59.11c.33.059 2.8.67 5.6 1.4 9.3 2.4 16 3.4 26 3.7 6.5.25 15 .022 20-.51 4.7-.57 11-1.8 14-2.6.79-.23 1.6-.41 1.8-.41.24 0 .61.5 1.1 1.5 5.7 11 16 22 27 28 .74.4 1.6.87 1.9 1.1.28.18 1 .52 1.6.76.97.38 1.1.39 1.6.14 1.9-1 13-12 17-17 4.7-5.5 7.8-11 9.5-16 .72-2.1.8-2.3 1.4-2.3.32-.037 2 .31 3.7.76 5.1 1.4 14 3 19 3.5 9.3 1 16 1.1 24 .4 5.8-.54 12-2.2 17-4.6 1.5-.74 2.8-1.3 2.8-1.3.043 0 .12 3.5.16 7.8.14 12 .7 27 1.2 30 .11.77.2 2.1.2 2.9 0 1.3.35 6.4.72 11 .062.68.2 2.3.31 3.5.11 1.2.29 3 .41 3.8.12.85.41 3.7.65 6.3.68 7.3.62 6.8 1.7 15 .17 1.2.39 3 .49 3.9.18 1.7.41 3.3.82 5.9.64 3.9.87 5.3 1.1 6.4.42 2.2.26 3.9-.94 9.9-.63 3.2-1.4 7-1.7 8.5-.53 2.8-1.7 8.1-2.3 10-.16.57-.44 1.7-.62 2.6-.18.85-.66 2.9-1.1 4.6-1.2 5.1-1.7 7.1-2.2 9.5-.29 1.2-1 4.3-1.6 6.7-.57 2.4-1.3 5.6-1.7 7.1-.83 3.6-1.3 5.2-1.9 7.3-.79 2.6-.77 2.7.53 2.7 5.4.11 28 .4 30 .38 1.9-.02 2.3.033 2.7.35.47.41 2.9 5.6 2.7 5.8-.17.17-2.5.076-3.7-.15-.57-.11-2.2-.34-3.7-.52s-3.3-.41-4.1-.52c-.79-.11-2.6-.3-4.1-.42-1.5-.12-3.6-.3-4.6-.39-2-.17-5-.42-9.8-.81-4.9-.39-12-.64-20-.63-9.1.014-15 .4-22 1.4-1.8.26-5.9 1.2-7.1 1.6-.46.16-.96.3-1.1.3-.21 0-.29-1.7-.37-8.4-.15-13-.6-24-1.2-29-.11-.91-.29-2.9-.41-4.4-.54-7.4-2.1-17-3.6-24-.93-3.8-1.1-4.6-2.1-7.4-2.1-6.2-4-10-5-10-.53 0-5 1.9-8.3 3.6-11 5.5-19 12-23 19-.38.6-.83 1.1-1 1.1-.19-.004-.75-.63-1.3-1.4-2.5-3.7-5.1-7-8.9-11-5.3-5.5-10-8.9-16-11-1.9-.73-6-1.5-6.3-1.1-.36.3-1.1 3.4-2.2 9.1-.46 2.4-.93 4.8-1 5.4-.17.83-.77 4.8-1.2 8.1-.094.68-.27 2-.4 2.9-.54 3.9-.63 4.7-.85 7.1-.12 1.4-.3 3.4-.4 4.3-.096.96-.33 5.4-.52 9.8-.3 6.9-.31 9-.11 15 .13 3.8.32 7.7.43 8.6.11.91.35 3 .53 4.6.18 1.6.51 4 .73 5.2.22 1.3.46 2.7.53 3.2.13.96.17.94-1.3.67zm175-9.2c-.065-.065-6-.21-13-.33-26-.41-29-.48-30-.66-.36-.23-.45.19 2.1-9.9 1.2-4.8 2.5-10 2.9-12s.99-4.3 1.3-5.6c.33-1.3.84-3.6 1.1-5 .29-1.4.69-3.3.9-4.2.21-.87.53-2.3.72-3.2 1.1-5.4 1.4-7 1.9-9.3 1.5-7.5 2.2-11 2.8-15 .34-2 .75-4.4.93-5.4.17-.96.51-3 .74-4.5 1.5-9.8 1.5-9.5 2.1-14 .17-1.2.55-4.2.82-6.7.28-2.5.6-5.2.71-6.1.11-.85.26-2.2.32-3.1.064-.85.2-2.1.31-2.9.35-2.5.94-11 1.3-19 .51-12 .15-23-1.1-33-.21-1.8-.2-2 .17-2.3.34-.28 3.7-1 11-2.4.91-.17 2.8-.45 4.2-.62 6.6-.81 7.8-.95 9.6-1.1 6-.61 22-.85 27-.4 3.7.31 6.3.57 9.1.92 1.4.17 3.3.4 4.2.51 2.2.26 11 2.2 15 3.2 12 3.3 22 8.1 31 14 1.4.95 2.9 2.1 3.4 2.5.5.41 1.4 1.1 2 1.6.62.47 1.4 1.2 1.8 1.6s1.1 1.3 1.6 1.8c.7.73 1.3 1.7 2.1 3.5.64 1.4 1.5 3.2 1.9 4.1 1 2.1 1.7 4.4 2.8 9.6.75 3.5 1 6.8 1 12-.005 6.4-.26 8.5-2 16-.48 2.1-2.3 7.2-3.5 10-1.6 3.7-4.7 9.3-6.3 11-2.7 3.6-7.4 8.3-11 11-5.2 3.8-9.7 5.7-16 7.2-2.5.54-3.4.64-7.1.75-6.5.18-9.7-.14-16-1.6-5.4-1.3-12-3.8-16-5.9-.76-.45-1.5-.83-1.7-.83-.19 0-.58.52-.94 1.2-2.4 4.7-5.1 15-6.1 23-.19 1.6-.45 3.8-.59 4.9-.34 2.8-.35 19-.004 21 .14.96.35 3 .48 4.4.12 1.5.35 3.4.51 4.2.16.85.35 2.2.42 2.9l.14 1.4h-9.8c-5.4 0-9.8-.054-9.9-.12zm37-97c10-1.1 16-4.2 18-9.1 1.4-3.6-.71-8.8-4.7-12-3.3-2.4-7.3-3.8-13-4.4-2.1-.24-8.1-.31-10-.11-3.2.3-5.1.58-5.2.79-.077.12-.099.46-.049.75.079.45.41 4.9 1.3 18 .11 1.6.27 3.6.35 4.4l.15 1.5 1.6.13c2.3.18 8.6.002 12-.34zm-521 95c0-.093.13-.63.3-1.2 1.6-5.5 3.6-15 4.8-23 2-14 3.1-27 3.6-43 .24-7.5.12-32-.19-38l-.13-2.4-7.3.056c-4 .031-18 .24-31 .46-13 .22-24 .36-24 .3-.11-.11.04-.58 1.7-5.3 4.3-12 7-25 8.2-40 .4-4.7.33-20-.12-24-.6-6.6-2-15-3.1-19-.5-1.7-.62-2.5-.38-2.5.25 0 2.6 1.3 4.3 2.4.62.4 1.4.83 1.7.97 1.1.5 5.9 3.1 6.8 3.7.5.34 1.9 1.1 3.2 1.6 1.2.55 2.5 1.2 2.8 1.4.65.46 3.8 1.9 4.1 1.9.58 0 8.2 2.4 9.2 3 .57.28 1.9.53 5.7 1 1.4.18 3.1.43 3.8.55.74.12 2 .29 2.9.38 5.7.62 8.1.85 11 1 1.6.11 4.5.3 6.4.43 4.6.31 30 .33 37 .02 2.8-.13 8.6-.48 13-.77 9.7-.66 9.5-.64 13-1 1.7-.17 3.5-.34 4-.39 1.8-.15 7.7-.75 9.4-.94 2.7-.31 6.8-.83 7.3-.92.28-.049 1.7-.23 3.2-.41 4-.48 7.4-.93 9.7-1.3 1.1-.17 2.9-.43 3.9-.57 1.7-.24 2.8-.41 8.3-1.3 3.3-.5 11-1.1 11-.89-3.5e-5.044-.33.99-.73 2.1-.67 1.9-.96 2.8-2 6.6-.64 2.4-.97 4.2-1.4 7.7-.81 6.7-.94 9.5-.66 14 .14 2.3.35 4.6.45 5.1.11.57.28 1.9.39 2.9.24 2.3.57 4 1.7 8.5 1.4 5.7 3.2 11 5.8 17 .57 1.3.98 2.4.9 2.5-.076.076-1.3-.19-2.8-.6-3.2-.91-4-1.1-6.4-1.5-1.1-.18-2.4-.42-2.9-.52-4.9-.99-19-.98-27 .024-4.9.7-8.3 1.3-12 2.2-4.2 1-5.1 1.2-8.1 2.2-4.2 1.3-4.8 1.6-5.1 2.2-1 2.6-1.2 23-.39 38 .2 3.6.99 12 1.5 15 .12.85.3 2.2.41 3.1.39 3 1 7.3 1.3 9.1.59 3.3.73 4 1 5.7.49 2.8 2.3 11 3.6 16 1.6 6.1 2.9 11 4.9 17 2 6.1 2.1 6.4 2 6.6-.072.072-1.8-.39-3.9-1-7-2.1-11-3.1-17-4.2-1-.2-4-.64-7-1-2.1-.28-13-.28-15-.003-6.1.75-12 1.8-17 3.4-3.7 1.1-9.9 3-12 3.6-.62.23-2.4.87-3.9 1.4-1.5.55-4 1.5-5.5 2-1.5.57-2.9 1-3.2 1-.23 0-.57.13-.75.3-.25.23-3.2 1.5-3.5 1.6-.028 4e-4-.051-.075-.051-.17zm702-109c-2.8-.24-5.6-.55-7.3-.83-.86-.14-1-.45-1.3-2.7-.13-1-.36-2.1-.5-2.4-.38-.75-2.1-6.4-3-10-.98-3.8-2.2-7.1-4.4-11-3-5.8-7-12-11-16-.46-.51-1.3-1.7-1.8-2.6-.54-.91-1.2-1.9-1.4-2.1-.66-.84-.49-1.2 1-2 4.1-2.4 11-4.8 18-6.4 1.9-.42 5.7-1 8.8-1.3 3.6-.39 14-.45 17-.1 1.1.12 2.6.27 3.3.33 3.4.29 11 2 15 3.5 8 2.7 14 5.9 18 10 4.3 4.2 6.2 7.5 7 12 .4 2.3.4 3.2.024 5.3-.48 2.7-1.6 5.4-3.2 7.8-.78 1.2-5.7 6.1-7 7-3.2 2.2-8 4.9-10 5.7-.43.15-2.1.76-3.6 1.3-3.1 1.1-9.8 2.9-12 3.3-4.5.66-6 .84-9.4 1.1-4.2.36-8.6.36-13 .004z" stroke-width="1.3"/>
+        <path d="m19 484c-.059-.095.17-1 .51-2.1.34-1.1 1-3.4 1.5-5.1.51-1.8 1.1-3.4 1.3-3.6.33-.37.66-.38 7.3-.28 6.4.1 10 .35 17 .98 1.2.12 2.3.27 2.5.34.17.062-.76.33-2.1.61-1.3.27-3.2.68-4.3.9-7 1.4-17 5-22 7.9-1 .59-1.5.71-1.7.41zm384-1.1c-.51-.093-2-.31-3.4-.49-1.4-.18-3.1-.41-3.8-.53-.74-.11-2.7-.35-4.3-.53-1.6-.18-3.6-.4-4.4-.5-.79-.096-2.5-.27-3.8-.39-1.3-.12-3.3-.31-4.5-.43-5.4-.52-11-.7-20-.71-11-.012-12 .072-19 .92-1.9.25-5.3.91-6.9 1.4-.61.17-1.2.23-1.3.14-.095-.095-.096-.63-.00044-1.2.16-.97 1.2-6.4 2.8-15 .15-.79.42-2.5.61-3.7.19-1.2.47-3 .64-3.8.53-2.8 1.5-8.8 1.7-10 .19-1.5.4-2.8 1.2-7.6.22-1.3.55-3.5.74-4.9.18-1.4.57-4.1.85-6 .28-1.9.61-4 .72-4.7.24-1.6 1-7.5 1.2-8.8.054-.51.17-1 .25-1.1.084-.09 1.1-.38 2.2-.65 2.2-.53 5.5-1.7 10-3.5 9.6-3.8 11-4.8 18-13 3.6-4.4 6.3-8.3 7.1-10 .28-.68.72-1.7.97-2.2.99-2 2.1-5.9 3.7-13 .31-1.4.8-3.3 1.1-4.3.94-3.4 1.3-6.6 1.3-13-.005-4.5-.1-6-.6-9.6-.31-2.2-1.5-7.4-2.1-9.5-.29-.91-.53-1.8-.54-2-.021-.78-1.8-5.4-2.6-6.5-1.1-1.7-8.6-9-11-11-1.3-.99-3.2-2.3-4.3-3-5.9-3.6-7-4.3-8.8-5-1.1-.42-2.3-.95-2.8-1.2-.45-.23-2.3-1-4-1.7-1.8-.71-3.3-1.4-3.4-1.6-.13-.17-.29-.74-.36-1.3-.071-.52-.55-2.6-1.1-4.7-.51-2-.93-3.9-.93-4.1 0-.22.21-.48.46-.59.83-.35 11-2.4 15-3 1.2-.18 2.5-.37 2.7-.43.52-.11 7.4-1.1 8.5-1.2.4-.046 2.1-.22 3.8-.4 1.7-.17 4-.41 5-.54 6.2-.71 20-1.1 27-.81 6.9.26 11 .52 13 .83.74.12 2.6.4 4.1.62 4.1.59 4.8.72 8.6 1.6 8 2 15 4.7 20 8.1 5.4 3.5 10 8.7 13 14 .49.91 1.3 2.1 1.9 2.8 2.4 2.6 5.6 7.2 8.8 13 3.5 5.9 6.9 15 8.1 22 .18.96.43 2.3.57 3 .36 1.8.34 12-.026 14-.91 5-1.2 6.1-2.3 9.6-3.2 9.7-11 21-18 25-7.7 5-16 8.1-26 9.1-3 .33-12 .32-15-.021-6.2-.7-13-2.2-17-4-.93-.36-1.7-.54-1.8-.44-.28.29-.49 3.7-1 17-.27 6.7-.27 18-.00086 24 .58 13 .88 17 1.2 19 .11.69.2 1.8.2 2.4 0 .66.091 1.8.2 2.5.11.69.31 2.1.43 3.1.13 1 .31 2.4.41 3 .6 3.9.78 5.2.68 5.4-.13.2-1.3.18-2.5-.05zm7.7-125c4.7-.56 8.5-1.7 11-3.2 1.3-.86 2.8-2.5 3.4-3.7.39-.85.45-1.3.43-3.6-.015-2.2-.098-2.9-.46-3.9-1.4-3.9-4.5-6.1-10-7.4-.85-.19-2.9-.27-6.5-.26-4.4.009-5.7.085-8 .46-3.4.56-9.3 1.9-9.6 2.1-.18.18-.017 2.9.81 13 .16 2 .3 4 .3 4.4.006.4.087.84.18.99.48.76 14 1.1 19 .54zm-332 119c-.98-.36-3.3-1-7.8-2.1-1.8-.46-3.8-.96-4.4-1.1-2.3-.6-8.8-1.9-11-2.3-1.2-.18-2.7-.41-3.3-.51-3.4-.62-4.2-.74-10-1.3-4.1-.44-15-.62-19-.32-6.1.44-13 1.8-17 3.6-1.1.42-1.4.48-1.6.27-.14-.14-.67-1.2-1.2-2.4-.5-1.2-1.4-3.2-2-4.5-.59-1.3-1.5-3.4-2.1-4.6-.54-1.2-1.5-3.4-2.2-4.8-.64-1.4-1.6-3.5-2.1-4.6-.49-1.1-1.7-3.9-2.8-6.2-1-2.3-1.9-4.2-1.9-4.3 0-.044-.32-.76-.7-1.6-.39-.83-1.5-3.4-2.5-5.6-1-2.3-2-4.3-2.1-4.4-.31-.31-.5-.3-2.6.12-2.2.45-3.1.59-7.7 1.2-1.3.18-3 .42-3.7.52-4.7.67-13 1-19 .88-3.9-.092-9.4-.54-11-.89-.51-.11-2.2-.48-3.8-.81-1.6-.33-3.7-.89-4.7-1.2-2.5-.84-2.7-.82-3.2.25-.99 2.1-3.3 9.1-4.1 13-1.6 6.5-2.1 8.7-2.8 14-.18 1.2-.42 2.7-.54 3.4-.28 1.7-.58 4.4-.8 7.4-.18 2.4-.33 3.2-.6 3.2-.073.00078-.97-.18-2-.41-2.9-.64-7.3-1.5-8.7-1.7-3.2-.43-7.1-.84-10-1.1-4.1-.31-21-.3-26 .005-8.3.49-11 .68-11 .78-.23.06-1.2.19-2.3.29-2.2.22-4.9.53-7.2.83-4.1.54-7.7 1-8.1 1.1-.24.067-.49.032-.56-.076-.11-.18.45-1.1 4.6-7.6.87-1.4 2.3-3.6 3.2-5 2.2-3.6 2-3.4 3.7-3.2 3.9.34 10 1.7 15 3.1.87.27 3.6 1.1 6.2 1.7 2.5.68 5.6 1.6 6.9 2 2.9.94 3.6 1.1 3.9.98.48-.19.32-1.2-.43-2.7-1.7-3.3-4.5-11-5.2-13-.18-.76-.55-2.1-.82-3-.27-.87-.88-3.1-1.3-5-.46-1.9-1-3.9-1.3-4.5-.35-.9-1.3-4.5-2.3-9.1-.16-.68-.5-2.2-.75-3.3l-.47-2 .41-.74c.23-.41 1.1-1.9 1.9-3.4 3.9-6.9 5.5-9.8 7.9-14 .7-1.3 1.5-2.8 1.8-3.3.28-.51 1.1-2 1.8-3.3.68-1.3 1.7-3.2 2.2-4.2 3.4-6.4 16-32 16-32 0-.053 1.1-2.3 2.3-5 1.3-2.7 2.8-6 3.4-7.3.59-1.3 1.2-2.4 1.3-2.5.54-.21 7.8.94 12 1.8 3.3.79 8.1 2.1 8.8 2.5l.67.31 1.3-1.2c.73-.65 1.3-1.2 1.3-1.3-.0055-.065-.38-.49-.83-.94-1.4-1.4-3.8-5.5-5.1-8.8-2-5-3.2-8.4-3.8-10-.81-3.1-2-9-2-9.8-.00036-.52 2.6-7.5 3.4-9 .14-.28.46-.54.71-.58.25-.036 1.9.32 3.7.79 13 3.3 22 5.3 31 6.5 1.4.18 3 .41 3.7.53 3.4.57 17 .56 19-.016.4-.12 1-.21 1.3-.21.68.003 2.6-.38 5.9-1.2 2.7-.67 6.8-2.2 9.9-3.6 1.3-.63 2.5-1.1 2.7-1.1.21 0 .43.39.64 1.2.18.65.68 2.4 1.1 3.9.44 1.5 1.4 4.8 2.2 7.3 1.5 5.1 2.5 8.3 5.3 17 1.1 3.3 2.2 6.8 2.5 7.9 2.3 7.3 9.3 27 13 35 .7 1.8 1.7 4.4 2.3 5.9 1.2 3.2 5 13 6.3 16 .53 1.2 1.4 3.3 2 4.6 1.1 2.5 5.8 13 8.3 18 4.7 10 13 25 17 33 4.8 8.1 6.7 11 12 18 .61.89 1.1 1.7 1.1 1.8 0 .27-1 .18-2-.16zm-114-81c2.6-.21 5.1-.62 5.1-.82 0-.07-.42-1.1-.93-2.3-2.5-5.7-4.6-14-6.2-26-.073-.54-.21-.98-.29-.98-.089 0-.42.63-.74 1.4-.88 2.1-3.5 7.3-4.8 9.5-2.6 4.5-6.2 9.4-9.5 13-1.1 1.3-2 2.4-2 2.5-.005.22.26.34 3 1.4 5.1 1.9 11 2.7 16 2.2zm499 81c-.13-.16.05-1.1.53-2.7 1.5-5 3.1-13 3.7-18 .19-1.6.43-3.4.53-4 .1-.57.24-1.9.3-2.9.062-1 .21-2.7.32-3.7.57-5 .57-21-.005-27-.12-1.2-.31-3.5-.43-4.9-.24-3-.46-5-.86-8.5-.3-2.5-.23-2.9.69-3.5 1.3-.84 8.2-8.7 10-11 1.8-2.6 4.6-8.6 5.8-13 .58-1.9 1.3-4.3 1.7-5.3.6-1.7.71-2.4 1.2-7.1.11-1.1.24-2 .31-2 .065-.065 1-.003 2.1.14 1.1.14 3 .38 4.3.54 8.8 1.1 19 1.1 29-.1 6.1-.68 14-2.4 19-4.3 1.7-.57 3.2-1 3.4-1 .48 0 .47.21-.14 2.5-.28 1.1-.88 3.8-1.3 6-1 5.1-1.4 7.6-1.9 13-.12 1.5-.31 3.6-.43 4.8-.26 2.8-.27 18-.005 20 .11.91.3 2.9.42 4.4.12 1.5.3 3.5.4 4.3.095.85.28 2.6.4 3.8.13 1.2.28 2.4.33 2.6.054.17.32 1.7.6 3.3.86 5.1 1.5 8.5 1.9 10 3.6 15 6.3 23 12 37 .46 1.1.84 2.1.84 2.2 0 .23-2.3.029-4.6-.4-.91-.17-2.7-.43-4-.59-1.3-.16-3.4-.44-4.7-.61-8.9-1.2-23-2.1-34-2.1-10-.004-23 .9-29 2.1-.96.18-2.4.46-3.3.61-.85.16-2.1.42-2.8.58-.68.16-1.9.45-2.7.63-.8.18-3.3.92-5.6 1.6-3.1.96-4.2 1.2-4.4 1.1zm-361-1.3c-.99-.18-5.9-.97-7.7-1.2-1.4-.21-4.3-.49-8.4-.83-1.3-.11-2.9-.26-3.5-.34l-1.1-.15-2.4-3.6c-1.3-2-2.6-4-2.8-4.5-.2-.51-.9-1.9-1.6-3.1-.66-1.2-1.7-3.2-2.3-4.4-1.1-2.3-9.4-19-12-25-.93-1.8-3.2-6.3-5.1-10-7-14-10-20-11-20-.37-.43-.85-1.5-3.9-9.2-1.2-2.9-1.9-4.9-1.9-5.4.0051-.4.17-2 .36-3.6.19-1.6.43-3.9.53-5.2.099-1.3.33-4 .52-6 .38-4 .55-6.2.89-11 .12-1.9.32-4.6.43-6 .45-5.3.67-19 .54-33-.072-7.6-.19-14-.26-15l-.13-1.2.59.11c.33.059 2.8.67 5.6 1.4 9.3 2.4 16 3.4 26 3.7 6.5.25 15 .022 20-.51 4.7-.57 11-1.8 14-2.6.79-.23 1.6-.41 1.8-.41.24 0 .61.5 1.1 1.5 5.7 11 16 22 27 28 .74.4 1.6.87 1.9 1.1.28.18 1 .52 1.6.76.97.38 1.1.39 1.6.14 1.9-1 13-12 17-17 4.7-5.5 7.8-11 9.5-16 .72-2.1.8-2.3 1.4-2.3.32-.037 2 .31 3.7.76 5.1 1.4 14 3 19 3.5 9.3 1 16 1.1 24 .4 5.8-.54 12-2.2 17-4.6 1.5-.74 2.8-1.3 2.8-1.3.043 0 .12 3.5.16 7.8.14 12 .7 27 1.2 30 .11.77.2 2.1.2 2.9 0 1.3.35 6.4.72 11 .062.68.2 2.3.31 3.5.11 1.2.29 3 .41 3.8.12.85.41 3.7.65 6.3.68 7.3.62 6.8 1.7 15 .17 1.2.39 3 .49 3.9.18 1.7.41 3.3.82 5.9.64 3.9.87 5.3 1.1 6.4.42 2.2.26 3.9-.94 9.9-.63 3.2-1.4 7-1.7 8.5-.53 2.8-1.7 8.1-2.3 10-.16.57-.44 1.7-.62 2.6-.18.85-.66 2.9-1.1 4.6-1.2 5.1-1.7 7.1-2.2 9.5-.29 1.2-1 4.3-1.6 6.7-.57 2.4-1.3 5.6-1.7 7.1-.83 3.6-1.3 5.2-1.9 7.3-.79 2.6-.77 2.7.53 2.7 5.4.11 28 .4 30 .38 1.9-.02 2.3.033 2.7.35.47.41 2.9 5.6 2.7 5.8-.17.17-2.5.076-3.7-.15-.57-.11-2.2-.34-3.7-.52s-3.3-.41-4.1-.52c-.79-.11-2.6-.3-4.1-.42-1.5-.12-3.6-.3-4.6-.39-2-.17-5-.42-9.8-.81-4.9-.39-12-.64-20-.63-9.1.014-15 .4-22 1.4-1.8.26-5.9 1.2-7.1 1.6-.46.16-.96.3-1.1.3-.21 0-.29-1.7-.37-8.4-.15-13-.6-24-1.2-29-.11-.91-.29-2.9-.41-4.4-.54-7.4-2.1-17-3.6-24-.93-3.8-1.1-4.6-2.1-7.4-2.1-6.2-4-10-5-10-.53 0-5 1.9-8.3 3.6-11 5.5-19 12-23 19-.38.6-.83 1.1-1 1.1-.19-.004-.75-.63-1.3-1.4-2.5-3.7-5.1-7-8.9-11-5.3-5.5-10-8.9-16-11-1.9-.73-6-1.5-6.3-1.1-.36.3-1.1 3.4-2.2 9.1-.46 2.4-.93 4.8-1 5.4-.17.83-.77 4.8-1.2 8.1-.094.68-.27 2-.4 2.9-.54 3.9-.63 4.7-.85 7.1-.12 1.4-.3 3.4-.4 4.3-.096.96-.33 5.4-.52 9.8-.3 6.9-.31 9-.11 15 .13 3.8.32 7.7.43 8.6.11.91.35 3 .53 4.6.18 1.6.51 4 .73 5.2.22 1.3.46 2.7.53 3.2.13.96.17.94-1.3.67zm175-9.2c-.065-.065-6-.21-13-.33-26-.41-29-.48-30-.66-.36-.23-.45.19 2.1-9.9 1.2-4.8 2.5-10 2.9-12s.99-4.3 1.3-5.6c.33-1.3.84-3.6 1.1-5 .29-1.4.69-3.3.9-4.2.21-.87.53-2.3.72-3.2 1.1-5.4 1.4-7 1.9-9.3 1.5-7.5 2.2-11 2.8-15 .34-2 .75-4.4.93-5.4.17-.96.51-3 .74-4.5 1.5-9.8 1.5-9.5 2.1-14 .17-1.2.55-4.2.82-6.7.28-2.5.6-5.2.71-6.1.11-.85.26-2.2.32-3.1.064-.85.2-2.1.31-2.9.35-2.5.94-11 1.3-19 .51-12 .15-23-1.1-33-.21-1.8-.2-2 .17-2.3.34-.28 3.7-1 11-2.4.91-.17 2.8-.45 4.2-.62 6.6-.81 7.8-.95 9.6-1.1 6-.61 22-.85 27-.4 3.7.31 6.3.57 9.1.92 1.4.17 3.3.4 4.2.51 2.2.26 11 2.2 15 3.2 12 3.3 22 8.1 31 14 1.4.95 2.9 2.1 3.4 2.5.5.41 1.4 1.1 2 1.6.62.47 1.4 1.2 1.8 1.6s1.1 1.3 1.6 1.8c.7.73 1.3 1.7 2.1 3.5.64 1.4 1.5 3.2 1.9 4.1 1 2.1 1.7 4.4 2.8 9.6.75 3.5 1 6.8 1 12-.005 6.4-.26 8.5-2 16-.48 2.1-2.3 7.2-3.5 10-1.6 3.7-4.7 9.3-6.3 11-2.7 3.6-7.4 8.3-11 11-5.2 3.8-9.7 5.7-16 7.2-2.5.54-3.4.64-7.1.75-6.5.18-9.7-.14-16-1.6-5.4-1.3-12-3.8-16-5.9-.76-.45-1.5-.83-1.7-.83-.19 0-.58.52-.94 1.2-2.4 4.7-5.1 15-6.1 23-.19 1.6-.45 3.8-.59 4.9-.34 2.8-.35 19-.004 21 .14.96.35 3 .48 4.4.12 1.5.35 3.4.51 4.2.16.85.35 2.2.42 2.9l.14 1.4h-9.8c-5.4 0-9.8-.054-9.9-.12zm37-97c10-1.1 16-4.2 18-9.1 1.4-3.6-.71-8.8-4.7-12-3.3-2.4-7.3-3.8-13-4.4-2.1-.24-8.1-.31-10-.11-3.2.3-5.1.58-5.2.79-.077.12-.099.46-.049.75.079.45.41 4.9 1.3 18 .11 1.6.27 3.6.35 4.4l.15 1.5 1.6.13c2.3.18 8.6.002 12-.34zm-521 95c0-.093.13-.63.3-1.2 1.6-5.5 3.6-15 4.8-23 2-14 3.1-27 3.6-43 .24-7.5.12-32-.19-38l-.13-2.4-7.3.056c-4 .031-18 .24-31 .46-13 .22-24 .36-24 .3-.11-.11.04-.58 1.7-5.3 4.3-12 7-25 8.2-40 .4-4.7.33-20-.12-24-.6-6.6-2-15-3.1-19-.5-1.7-.62-2.5-.38-2.5.25 0 2.6 1.3 4.3 2.4.62.4 1.4.83 1.7.97 1.1.5 5.9 3.1 6.8 3.7.5.34 1.9 1.1 3.2 1.6 1.2.55 2.5 1.2 2.8 1.4.65.46 3.8 1.9 4.1 1.9.58 0 8.2 2.4 9.2 3 .57.28 1.9.53 5.7 1 1.4.18 3.1.43 3.8.55.74.12 2 .29 2.9.38 5.7.62 8.1.85 11 1 1.6.11 4.5.3 6.4.43 4.6.31 30 .33 37 .02 2.8-.13 8.6-.48 13-.77 9.7-.66 9.5-.64 13-1 1.7-.17 3.5-.34 4-.39 1.8-.15 7.7-.75 9.4-.94 2.7-.31 6.8-.83 7.3-.92.28-.049 1.7-.23 3.2-.41 4-.48 7.4-.93 9.7-1.3 1.1-.17 2.9-.43 3.9-.57 1.7-.24 2.8-.41 8.3-1.3 3.3-.5 11-1.1 11-.89-3.5e-5.044-.33.99-.73 2.1-.67 1.9-.96 2.8-2 6.6-.64 2.4-.97 4.2-1.4 7.7-.81 6.7-.94 9.5-.66 14 .14 2.3.35 4.6.45 5.1.11.57.28 1.9.39 2.9.24 2.3.57 4 1.7 8.5 1.4 5.7 3.2 11 5.8 17 .57 1.3.98 2.4.9 2.5-.076.076-1.3-.19-2.8-.6-3.2-.91-4-1.1-6.4-1.5-1.1-.18-2.4-.42-2.9-.52-4.9-.99-19-.98-27 .024-4.9.7-8.3 1.3-12 2.2-4.2 1-5.1 1.2-8.1 2.2-4.2 1.3-4.8 1.6-5.1 2.2-1 2.6-1.2 23-.39 38 .2 3.6.99 12 1.5 15 .12.85.3 2.2.41 3.1.39 3 1 7.3 1.3 9.1.59 3.3.73 4 1 5.7.49 2.8 2.3 11 3.6 16 1.6 6.1 2.9 11 4.9 17 2 6.1 2.1 6.4 2 6.6-.072.072-1.8-.39-3.9-1-7-2.1-11-3.1-17-4.2-1-.2-4-.64-7-1-2.1-.28-13-.28-15-.003-6.1.75-12 1.8-17 3.4-3.7 1.1-9.9 3-12 3.6-.62.23-2.4.87-3.9 1.4-1.5.55-4 1.5-5.5 2-1.5.57-2.9 1-3.2 1-.23 0-.57.13-.75.3-.25.23-3.2 1.5-3.5 1.6-.028 4e-4-.051-.075-.051-.17zm702-109c-2.8-.24-5.6-.55-7.3-.83-.86-.14-1-.45-1.3-2.7-.13-1-.36-2.1-.5-2.4-.38-.75-2.1-6.4-3-10-.98-3.8-2.2-7.1-4.4-11-3-5.8-7-12-11-16-.46-.51-1.3-1.7-1.8-2.6-.54-.91-1.2-1.9-1.4-2.1-.66-.84-.49-1.2 1-2 4.1-2.4 11-4.8 18-6.4 1.9-.42 5.7-1 8.8-1.3 3.6-.39 14-.45 17-.1 1.1.12 2.6.27 3.3.33 3.4.29 11 2 15 3.5 8 2.7 14 5.9 18 10 4.3 4.2 6.2 7.5 7 12 .4 2.3.4 3.2.024 5.3-.48 2.7-1.6 5.4-3.2 7.8-.78 1.2-5.7 6.1-7 7-3.2 2.2-8 4.9-10 5.7-.43.15-2.1.76-3.6 1.3-3.1 1.1-9.8 2.9-12 3.3-4.5.66-6 .84-9.4 1.1-4.2.36-8.6.36-13 .004z" strokeWidth="1.3"/>
         </g>
     )
   },
@@ -258,7 +258,7 @@ const friendsInfo = [
     title: 'Huld',
     url: 'https://huld.io/',
     logo: (
-      <g className="f1" transform="matrix(.59 0 0 .59 153 -145)" stroke-width="2">
+      <g className="f1" transform="matrix(.59 0 0 .59 153 -145)" strokeWidth="2">
       <path d="m226 246h-24c-2.4 0-4.4 2-4.4 4.4v242c0 2.4 2 4.4 4.4 4.4h24c2.4 0 4.4-2 4.4-4.4v-242c0-2.4-2-4.4-4.4-4.4z"/>
       <path d="m426 246h-24c-2.4 0-4.4 2-4.4 4.4v94c-14-11-30-17-48-17-15 0-28 3.6-40 11-26 16-42 44-42 76s15 60 42 76c11 6.8 26 11 40 11 19 0 36-6.4 48-18v10c0 2.4 2 4.4 4.4 4.4h24c2.4 0 4.4-2 4.4-4.4v-242c0-2.4-2-4.4-4.4-4.4zm-30 200c-11 14-26 22-42 22-30 0-52-24-52-54s22-54 50-54c17 0 34 8.2 44 22 .4.52.64 1.2.64 1.9v60c-.02.68-.24 1.3-.62 1.9z"/>
       <path d="m150 330h-24c-2.4 0-4.4 2-4.4 4.4v80c0 22-4.6 32-9.4 38-7.6 10-20 16-34 16-13 0-24-5.2-32-14-8.6-11-9-24-9.2-40l-.02-1.1c-.26-32-16-60-42-76-11-7-26-11-40-11-22 0-44 9.6-58 26v-104c0-2.4-2-4.4-4.4-4.4h-24c-2.4 0-4.4 2-4.4 4.4v242c0 2.4 2 4.4 4.4 4.4h24c2.4 0 4.4-2 4.4-4.4v-88c0-.16.02-.34.04-.5 3.2-22 26-46 54-46 32 0 52 28 52 54l.1 3.2c.5 18 1.1 38 15 56 14 18 36 28 60 28s46-11 62-30c11-14 16-32 16-60v-78c.02-2.4-2-4.4-4.4-4.4z"/>
